Make polkadot-parachain compile for WASM
