+++ parachain/Cargo.toml
@@ -12,10 +12,12 @@ derive_more = { version = "0.14", optional = true }
 serde = { version = "1.0", default-features = false, features = [ "derive" ] }
 
 rstd = { package = "sr-std", git = "https://github.com/paritytech/substrate", branch = "polkadot-master", default-features = false }
-shared_memory = { version = "0.8", optional = true }
 lazy_static = { version = "1.3.0", optional = true }
 parking_lot = { version = "0.7.1", optional = true }
 log = { version = "0.4.6", optional = true }
+
+[target.'cfg(not(target_os = "unknown"))'.dependencies]
+shared_memory = { version = "0.8", optional = true }
 
 [dev-dependencies]
 tiny-keccak = "1.4"
